replace curve shapes when mirroring controls

+++ src/pulse/scripts/pulse/colors.py
@@ -1,4 +1,3 @@
-
 __all__ = [
     'hexToRGB01',
     'RGB01ToHex',
@@ -9,7 +8,7 @@ def RGB01ToHex(rgb):
     """
     Return an RGB color value as a hex color string.
     """
-    return '#%02x%02x%02x' % tuple([x * 255 for x in rgb])
+    return '#%02x%02x%02x' % tuple([int(x * 255) for x in rgb])
 
 
 def hexToRGB01(hexColor):
@@ -17,4 +16,4 @@ def hexToRGB01(hexColor):
     Return a hex color string as an RGB tuple of floats in the range 0..1
     """
     h = hexColor.lstrip('#')
-    return tuple([x/255.0 for x in [int(h[i:i+2], 16) for i in (0, 2, 4)]])
+    return tuple([x / 255.0 for x in [int(h[i:i + 2], 16) for i in (0, 2, 4)]])
